cipher-event-logger/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bson==0.5.10
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-restx==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
pika==1.2.0
py-healthcheck==1.10.1
pyaml==21.8.3
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
Werkzeug==2.0.1
